+++ include/tokamak/particle_push.hpp
@@ -8,8 +8,8 @@ Vec3 BorisVelocityStep(
     const Vec3& velocity,
     const Vec3& electricField_VPerM,
     const Vec3& magneticField_T,
-    float chargeToMass_CPerKg,
-    float dt_s);
+    double chargeToMass_CPerKg,
+    double dt_s);
 
 bool ReflectAtTokamakWall(
     Vec3* position,
